refactor(desktop): move querySurfaceChoice to its own file (line-count ratchet)
The pure, unit-tested onboarding-isolation surface decision lives in a small
ChatProvider extension file instead of growing the already-oversized
ChatProvider.swift past its frozen line-count baseline.

Failure-Class: none
