+++ desktop/macos/e2e/flows/chat-hermetic.yaml
@@ -20,6 +20,7 @@ covers:
   - desktop/macos/Desktop/Sources/MainWindow/Pages/ChatPage.swift
   - desktop/macos/Desktop/Sources/Providers/ChatProvider.swift
   - desktop/macos/Desktop/Sources/Providers/ChatProvider+JournalProjection.swift
+  - desktop/macos/Desktop/Sources/Providers/ChatProvider+QuerySurface.swift
   - desktop/macos/Desktop/Sources/Providers/ChatProvider+Skills.swift
   - desktop/macos/Desktop/Sources/Providers/ChatSkillCatalog.swift
   - desktop/macos/Desktop/Sources/Providers/ChatProvider+HarnessReset.swift
